windows-emulator: prevent uint32 wrap in cmd_begin_rendering attachment count

+++ src/windows-emulator/devices/gpu_bridge.cpp
@@ -2759,8 +2759,8 @@ namespace sogen
                     {
                         return vk_error_initialization_failed;
                     }
-                    const uint32_t total = req.color_attachment_count + (req.has_depth ? 1u : 0u) + (req.has_stencil ? 1u : 0u);
-                    if (static_cast<size_t>(total) * sizeof(gpu_bridge::rendering_attachment) > size - sizeof(req))
+                    const size_t total = static_cast<size_t>(req.color_attachment_count) + (req.has_depth ? 1u : 0u) + (req.has_stencil ? 1u : 0u);
+                    if (total * sizeof(gpu_bridge::rendering_attachment) > size - sizeof(req))
                     {
                         return vk_error_initialization_failed;
                     }
